adding new features

+++ tests/test_indicators_comprehensive.py
@@ -11,17 +11,17 @@ def test_ema_and_sma_calculations():
 
 def test_rsi_calculation():
     # Strong uptrend prices
-    prices = [100 + i * 2 for i in range(30)]
+    prices = [100.0 + (i * 2.0) for i in range(30)]
     rsi = TechnicalAnalysisService.calculate_rsi(prices, 14)
     assert rsi > 70.0 # Overbought in strong uptrend
 
     # Strong downtrend prices
-    prices_down = [200 - i * 2 for i in range(30)]
+    prices_down = [200.0 - (i * 2.0) for i in range(30)]
     rsi_down = TechnicalAnalysisService.calculate_rsi(prices_down, 14)
     assert rsi_down < 30.0 # Oversold in strong downtrend
 
 def test_macd_calculation():
-    prices = [100 + i * 1.5 for i in range(50)]
+    prices = [100.0 + (i * 1.5) for i in range(50)]
     macd = TechnicalAnalysisService.calculate_macd(prices)
     assert hasattr(macd, "value")
     assert hasattr(macd, "signal")
@@ -29,7 +29,7 @@ def test_macd_calculation():
 
 def test_atr_and_adx_calculations():
     candles = [
-        {"open": 100 + i, "high": 105 + i, "low": 98 + i, "close": 102 + i, "volume": 1000 + i * 10}
+        {"open": 100.0 + float(i), "high": 105.0 + float(i), "low": 98.0 + float(i), "close": 102.0 + float(i), "volume": 1000.0 + float(i * 10)}
         for i in range(40)
     ]
     atr = TechnicalAnalysisService.calculate_atr(candles, 14)
@@ -39,19 +39,19 @@ def test_atr_and_adx_calculations():
     assert adx >= 0 and adx <= 100
 
 def test_bollinger_bands():
-    prices = [100.0 + (i % 5) for i in range(30)]
+    prices = [100.0 + float(i % 5) for i in range(30)]
     bb = TechnicalAnalysisService.calculate_bollinger_bands(prices, 20)
     assert bb.upper > bb.middle
     assert bb.middle > bb.lower
 
 def test_volume_metrics():
     candles = [
-        {"open": 100, "high": 105, "low": 98, "close": 102, "volume": 1000}
+        {"open": 100.0, "high": 105.0, "low": 98.0, "close": 102.0, "volume": 1000.0}
         for _ in range(25)
     ]
     # Add a massive volume spike on the latest candle
-    candles[-1]["volume"] = 3500
-    candles[-1]["close"] = 110 # Bullish breakout bar
+    candles[-1]["volume"] = 3500.0
+    candles[-1]["close"] = 110.0 # Bullish breakout bar
 
     vol = TechnicalAnalysisService.calculate_volume_metrics(candles)
     assert vol.is_volume_confirmed == True
